+++ eth_defi/revert_reason.py
@@ -8,11 +8,11 @@ Further reading
 import logging
 from typing import Union
 
-from eth_defi.abi import get_transaction_data_field
 from eth_tester.exceptions import TransactionFailed
 from hexbytes import HexBytes
 from web3 import Web3
 
+from eth_defi.abi import get_transaction_data_field
 
 logger = logging.getLogger(__name__)